app/app/requirements.txt
git+https://github.com/fullmakeralchemist/pytube/
git+https://github.com/gatagat/lap@new-packaging
streamlit
streamlit-webrtc
torch
ultralytics
pafy

altair
attrs
backports.zoneinfo
blinker
cachetools
certifi
charset-normalizer
click
cmake
contourpy
cycler
Cython
filelock
fonttools
gitdb
GitPython
idna
importlib-metadata
importlib-resources
Jinja2
jsonschema
jsonschema-specifications
kiwisolver
lapx
lit
markdown-it-py
MarkupSafe
matplotlib
mdurl
mpmath
networkx
numpy
nvidia-cublas-cu11
nvidia-cuda-cupti-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-cudnn-cu11
nvidia-cufft-cu11
nvidia-curand-cu11
nvidia-cusolver-cu11
nvidia-cusparse-cu11
nvidia-nccl-cu11
nvidia-nvtx-cu11
opencv-python-headless
packaging
pandas
Pillow
pkgutil-resolve-name
protobuf
psutil
py-cpuinfo
pyarrow
pydeck
Pygments
Pympler
pyparsing
python-dateutil
pytube
pytz
pytz-deprecation-shim
PyYAML
referencing
requests
rich
rpds-py
scipy
seaborn
six
smmap
streamlit
sympy
tenacity
toml
toolz
torch
torchvision
tornado
tqdm
triton
typing-extensions
tzdata
tzlocal
ultralytics
urllib3
validators
watchdog
zipp
